# docs/openapi.yaml
openapi: '3.0.3'
info:
  title: Superagent API
  description: REST API for the Superagent agent orchestration platform
  version: '1.0.0'
  contact:
    name: Superagent Team
servers:
  - url: http://localhost:3000/api/v1
    description: Local development
paths:
  /agents:
    get:
      summary: List all registered agents
      operationId: listAgents
      tags: [Agents]
      responses:
        '200':
          description: Agent list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Agent'
    post:
      summary: Register a new agent
      operationId: registerAgent
      tags: [Agents]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentInput'
      responses:
        '201':
          description: Agent registered
        '400':
          description: Invalid input
  /agents/{agentId}:
    get:
      summary: Get agent details
      operationId: getAgent
      tags: [Agents]
      parameters:
        - name: agentId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Agent details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Agent'
        '404':
          description: Agent not found
...
components:
  schemas:
    Agent:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        provider:
          type: string
        model:
          type: string
        status:
          type: string
          enum: [active, inactive, error]
        createdAt:
          type: string
          format: date-time
    AgentInput:
      type: object
      required: [name, provider, model]
      properties:
        name:
          type: string
        provider:
          type: string
        model:
          type: string
        config:
          type: object
...